+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed build error due to breaking change in `xmrs`.
+
 ## [0.21.1] - 2024/10/02
 
 ### Added