+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Optional serde support for agb-hashmap via the `serde` feature flag
+- Added `set_background_palette` to be able to set a single background palette.
 
 ### Fixed