+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Optional serde support for agb-hashmap via the `serde` feature flag
+- Optional serde support for agb-hashmap and agb-fixnum via the `serde` feature flag. The flag is also exposed in the agb crate.
 - Added `set_background_palette` to be able to set a single background palette.
 
 ### Changed